server_code/requirements.txt
matplotlib
pandas
anvil-uplink
python-docx